+++ src/os2datascanner/projects/static/js/index.js
@@ -1,27 +1,28 @@
 import '../scss/master.scss';
 
-$(document).ready(function(){
-  $("a[data-modal='modal:open']").click(function(e){
-      e.preventDefault();
+htmx.onLoad(function (content) {
 
-      // Find the target element (same as our href)
-      var target = $(this).attr("href");
+  $(content).find("a[data-modal='modal:open']").click(function (e) {
+    e.preventDefault();
 
-      // Find the src to set
-      var src = $(this).attr("data-src");
+    // Find the target element (same as our href)
+    const target = $(this).attr("href");
 
-      // Find the iframe in our target and set its src
-      $(target).find("iframe").attr("src",src);
+    // Find the src to set
+    const src = $(this).attr("data-src");
+
+    // Find the iframe in our target and set its src
+    $(target).find("iframe").attr("src", src);
   });
 
   // Toggle visiblity of expandable rows, start
-  $(document).on("click", "[data-toggle]", function() {
-    var expandTarget = $(this).attr("data-toggle");
+  $(document).on("click", "[data-toggle]", function () {
+    let expandTarget = $(this).attr("data-toggle");
 
     if ($(expandTarget).is('[hidden]')) {
       $(expandTarget).removeAttr('hidden');
     } else {
-      $(expandTarget).attr('hidden','');
+      $(expandTarget).attr('hidden', '');
     }
   });
   // Toggle visiblity of expandable rows, stop